net: add tcp_brutal congestion control
TCP Brutal is Hysteria's congestion control algorithm ported to TCP,
as a Linux kernel module. Information about Brutal itself can be found
in the Hysteria documentation. As an official subproject of Hysteria,
TCP Brutal is actively maintained to be in sync with the
Brutal implementation in Hysteria.

The code is from: https://github.com/apernet/tcp-brutal